Agrego dtype

+++ dataset/dataset.py
@@ -43,9 +43,9 @@ class MUSDB18Dataset(Dataset):
         x_audio = track.audio.T
         y_audio = track.targets[self.target].audio.T
         x = torch.as_tensor([librosa.stft(x_audio[0], n_fft=self.n_fft, hop_length=self.hop),
-                             librosa.stft(x_audio[1], n_fft=self.n_fft, hop_length=self.hop)])
+                             librosa.stft(x_audio[1], n_fft=self.n_fft, hop_length=self.hop)], dtype=torch.float)
         y = torch.as_tensor([librosa.stft(y_audio[0], n_fft=self.n_fft, hop_length=self.hop),
-                             librosa.stft(y_audio[1], n_fft=self.n_fft, hop_length=self.hop)])
+                             librosa.stft(y_audio[1], n_fft=self.n_fft, hop_length=self.hop)], dtype=torch.float)
         return x, y
 
     def __len__(self) -> int:
